implement object traversal for wmi object, make namespace and filter optional
